Adding a QD stock with the default buttpad length
For those of use who want to use their existing buttpad or refer a
smaller stock, I took the intersection of the bottom of the default
Lynx stock with the Ergo stock, then combined it with the upper QD
equipped body.

This was all performed on a copy of the Ergo stock in the step file, and
I exported a STL for convenience.